+++ src/richie/apps/courses/models/category.py
@@ -71,7 +71,10 @@ class Category(EsIdMixin, BasePageExtension):
         """
         return self.get_reverse_related_page_extensions(
             "course", language=language, include_descendants=include_descendants
-        ).filter(extended_object__node__parent__cms_pages__course__isnull=True)
+        ).filter(
+            extended_object__node__parent__cms_pages__course__isnull=True,
+            is_listed=True,
+        )
 
     def get_blogposts(self, language=None, include_descendants=True):
         """